ml-server/requirements.txt
flask>=2.2.5
flask-cors==3.0.10
tensorflow==2.16.1
numpy>=1.26.0
opencv-python==4.6.0.66
pillow>=10.0.0